+++ examples/miniapps/application-multiple-containers-runtime-overriding/example/__main__.py
@@ -1,0 +1,12 @@
+"""Main module."""
+
+from .containers import Application
+
+
+if __name__ == '__main__':
+    application = Application()
+    config = application.service.config()
+    config.build()
+
+    print(application.repository.site())
+    # Output: Adapter=[DB Path=[~/test]], queue=[Some storage]