+++ examples/miniapps/application-multiple-containers-runtime-overriding/example/services.py
@@ -1,0 +1,9 @@
+"""Services module."""
+
+
+class ConfigService:
+    def __init__(self, config):
+        self._config = config
+
+    def build(self):
+        self._config.from_dict({'default': {'db_path': '~/test'}})