+++ docs/main/changelog.rst
@@ -7,6 +7,14 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+4.25.0
+------
+- Add ``application-multiple-containers-runtime-overriding`` example. This example demonstrates
+  how to build application from multiple containers and override one container config from
+  another one in the runtime.
+  See issue: `#207 <https://github.com/ets-labs/python-dependency-injector/issues/207>`_.
+- Add attributes forwarding for the ``Dependency`` provider.
+
 4.24.0
 ------
 - Add docs on ``@containers.copy()`` decorator.